Reconfigure auth to be separate and before the app
Also utilize auth0's tenant sessions for logging in without a login screen.
